Squashed commit of the following:

commit 01629446c6a775864a9af1d0d981f57062c46aa1
Author: DEXREQ Automation <noreply+dexreq@oracle.com>
Date:   Fri Feb 14 20:34:10 2020 +0000

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 14 Feb 2020 20:34:10 +0000
    Running generation for: [[DEXREQ-1039, DEXREQ-1034, DEXREQ-1036, DEXREQ-1035, DEXREQ-1033, DEXREQ-1038]]

    Author: DEXREQ Automation <noreply+dexreq@oracle.com>
    Date:  Fri, 14 Feb 2020 20:09:50 +0000
    Updating pom.xml for [[DEXREQ-1039, DEXREQ-1034, DEXREQ-1036, DEXREQ-1035, DEXREQ-1033, DEXREQ-1038]]

    (automatically combined)

+++ database/vm_cluster.go
@@ -55,7 +55,7 @@ type VmCluster struct {
 	CpusEnabled *int `mandatory:"false" json:"cpusEnabled"`
 
 	// Size, in terabytes, of the DATA disk group.
-	DataStorageSizeInTBs *int `mandatory:"false" json:"dataStorageSizeInTBs"`
+	DataStorageSizeInTBs *float64 `mandatory:"false" json:"dataStorageSizeInTBs"`
 
 	// The shape of the Exadata infrastructure. The shape determines the amount of CPU, storage, and memory resources allocated to the instance.
 	Shape *string `mandatory:"false" json:"shape"`
